fix: update utils/i18n to include explicit namespace - remove setDefaultNamespace
