+++ src/types/i18n.d.ts
@@ -1,5 +1,4 @@
 declare module "@dhis2/d2-i18n" {
     export function t(value: string, namespace?: object): string;
     export function changeLanguage(locale: string);
-    export function setDefaultNamespace(namespace: string);
 }
