+++ src/webapp/contexts/app-context.ts
@@ -2,7 +2,6 @@ import React, { useContext } from "react";
 import { CompositionRoot } from "../../CompositionRoot";
 import { User } from "../../domain/entities/User";
 import { D2Api } from "../../types/d2-api";
-import i18n from "../../utils/i18n";
 
 export interface AppContextState {
     api: D2Api;
@@ -15,7 +14,6 @@ export const AppContext = React.createContext<AppContextState | null>(null);
 
 export function useAppContext() {
     const context = useContext(AppContext);
-    i18n.setDefaultNamespace("d2-multi-visualization");
     if (context) {
         return context;
     } else {
