+++ src/webapp/pages/app/Dhis2App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import i18n from "@dhis2/d2-i18n";
+import i18n from "../../../utils/i18n";
 import { Provider } from "@dhis2/app-runtime";
 import { D2Api } from "../../../types/d2-api";
 import App from "./App";
